requirements.txt
django
djangorestframework
environs
convertdate==2.1.3
cycler==0.10.0
Cython==0.29.6
ephem==3.7.6.0
fbprophet==0.4.post2
holidays==0.9.10
kiwisolver==1.0.1
lunardate==0.2.0
matplotlib==3.0.3
numpy==1.16.2
pandas==0.24.2
pyparsing==2.3.1
pystan==2.18.1.0
python-dateutil==2.8.0
pytz==2018.9
recurse==1.3.0
setuptools-git==1.2
six==1.12.0
tabula==1.0.5
tornado==6.0.1